requirements.txt
# local package
-e .

# external requirements
click~=7.0
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
tensorflow==2.8.0
minio~=7.1.13
setuptools~=65.6.3